Add missing type/enum exports
