+++ src/beatmapsets/search.ts
@@ -1,4 +1,4 @@
-import type { Beatmapset } from "../types/beatmap"
+import type { BeatmapsetSearchResult } from "../types/beatmap"
 import type { OAuthAccessToken } from "../types/oauth_access_token"
 import type { Fetch } from "../types/fetch"
 
@@ -7,10 +7,6 @@ import { urlParameterGenerator } from "../helpers/url_parameter_generator"
 import { OsuApiV2WebRequestError } from "../helpers/custom_errors"
 
 declare const fetch: Fetch
-
-export interface BeatmapsetSearchResult {
-    beatmapsets: Beatmapset[]
-}
 
 export const search = async (
     oauthAccessToken: OAuthAccessToken,
