+++ src/users/recentActivity.ts
@@ -1,67 +1,17 @@
 import type { OAuthAccessToken } from "../types/oauth_access_token"
+import type {
+    Event,
+    EventAchievement,
+    EventRank,
+    EventRankLost,
+} from "../types/event"
 import type { Fetch } from "../types/fetch"
-import type { Timestamp } from ".."
 
 import { baseUrlApiV2 } from "../types/api_info"
 import { urlParameterGenerator } from "../helpers/url_parameter_generator"
 import { OsuApiV2WebRequestError } from "../helpers/custom_errors"
-import { GameMode } from ".."
 
 declare const fetch: Fetch
-
-export enum EventType {
-    Achievement = "achievement",
-    BeatmapPlaycount = "beatmapPlaycount",
-    BeatmapsetApprove = "beatmapsetApprove",
-    BeatmapsetDelete = "beatmapsetDelete",
-    BeatmapsetRevive = "beatmapsetRevive",
-    BeatmapsetUpdate = "beatmapsetUpdate",
-    BeatmapsetUpload = "beatmapsetUpload",
-    Rank = "rank",
-    RankLost = "rankLost",
-    UserSupportAgain = "userSupportAgain",
-    UserSupportFirst = "userSupportFirst",
-    UserSupportGift = "userSupportGift",
-    UsernameChange = "usernameChange",
-}
-
-export interface Event {
-    created_at: Timestamp
-    id: number
-    type: EventType
-}
-
-export interface EventUser {
-    username: string
-    url: string
-}
-
-export interface EventBeatmap {
-    title: string
-    url: string
-}
-
-export interface EventAchievement extends Event {
-    type: EventType.RankLost
-    achievement: string // ?????
-    user: EventUser
-}
-
-export interface EventRank extends Event {
-    type: EventType.Rank
-    scoreRank: string
-    rank: number
-    mode: GameMode
-    beatmap: EventBeatmap
-    user: EventUser
-}
-
-export interface EventRankLost extends Event {
-    type: EventType.RankLost
-    mode: GameMode
-    beatmap: EventBeatmap
-    user: EventUser
-}
 
 export const recentActivity = async (
     oauthAccessToken: OAuthAccessToken,
